chore: 14355 Moved `State` class to platform core test fixtures
