+++ platform-sdk/platform-apps/demos/CryptocurrencyDemo/src/main/java/module-info.java
@@ -1,6 +1,7 @@
 module com.swirlds.demo.crypto {
     requires com.swirlds.common;
     requires com.swirlds.metrics.api;
+    requires com.swirlds.platform.core.test.fixtures;
     requires com.swirlds.platform.core;
     requires com.hedera.pbj.runtime;
     requires java.desktop;
