+++ platform-sdk/platform-apps/demos/HelloSwirldDemo/src/main/java/module-info.java
@@ -1,5 +1,6 @@
 module com.swirlds.demo.hello {
     requires com.swirlds.common;
+    requires com.swirlds.platform.core.test.fixtures;
     requires com.swirlds.platform.core;
     requires com.hedera.pbj.runtime;
     requires java.desktop;
