+++ platform-sdk/platform-apps/demos/StatsDemo/src/main/java/module-info.java
@@ -2,6 +2,7 @@ module com.swirlds.demo.stats {
     requires com.swirlds.base;
     requires com.swirlds.common;
     requires com.swirlds.metrics.api;
+    requires com.swirlds.platform.core.test.fixtures;
     requires com.swirlds.platform.core;
     requires java.desktop;
     requires static com.github.spotbugs.annotations;
