+++ platform-sdk/platform-apps/tests/ConsistencyTestingTool/build.gradle.kts
@@ -23,7 +23,6 @@ mainModuleInfo { annotationProcessor("com.swirlds.config.processor") }
 testModuleInfo {
     requires("com.hedera.node.hapi")
     requires("com.swirlds.common.test.fixtures")
-    requires("com.swirlds.platform.core.test.fixtures")
     requires("org.junit.jupiter.api")
     requires("org.mockito")
 }
