+++ platform-sdk/platform-apps/tests/ConsistencyTestingTool/src/main/java/module-info.java
@@ -3,6 +3,7 @@ module com.swirlds.demo.consistency {
     requires com.swirlds.common;
     requires com.swirlds.config.api;
     requires com.swirlds.logging;
+    requires com.swirlds.platform.core.test.fixtures;
     requires com.swirlds.platform.core;
     requires com.hedera.pbj.runtime;
     requires org.apache.logging.log4j;
