+++ platform-sdk/platform-apps/tests/MigrationTestingTool/src/main/java/module-info.java
@@ -6,6 +6,7 @@ module com.swirlds.demo.migration {
     requires com.swirlds.merkle;
     requires com.swirlds.merkledb;
     requires com.swirlds.metrics.api;
+    requires com.swirlds.platform.core.test.fixtures;
     requires com.swirlds.platform.core;
     requires com.swirlds.virtualmap;
     requires com.hedera.pbj.runtime;
