+++ platform-sdk/platform-apps/tests/PlatformTestingTool/src/main/java/module-info.java
@@ -33,6 +33,7 @@ module com.swirlds.demo.platform {
     requires com.swirlds.merkle;
     requires com.swirlds.merkledb;
     requires com.swirlds.metrics.api;
+    requires com.swirlds.platform.core.test.fixtures;
     requires com.swirlds.platform.core;
     requires com.swirlds.virtualmap;
     requires com.fasterxml.jackson.annotation;
