+++ platform-sdk/platform-apps/tests/StatsSigningTestingTool/src/main/java/module-info.java
@@ -3,6 +3,7 @@ module com.swirlds.demo.stats.signing {
     requires com.swirlds.common;
     requires com.swirlds.logging;
     requires com.swirlds.metrics.api;
+    requires com.swirlds.platform.core.test.fixtures;
     requires com.swirlds.platform.core;
     requires com.hedera.pbj.runtime;
     requires lazysodium.java;
