+++ platform-sdk/platform-apps/tests/StressTestingTool/src/main/java/module-info.java
@@ -4,6 +4,7 @@ module com.swirlds.demo.stress {
     requires com.swirlds.config.api;
     requires com.swirlds.logging;
     requires com.swirlds.metrics.api;
+    requires com.swirlds.platform.core.test.fixtures;
     requires com.swirlds.platform.core;
     requires com.hedera.pbj.runtime;
     requires org.apache.logging.log4j;
