+++ platform-sdk/swirlds-platform-core/src/testFixtures/java/module-info.java
@@ -7,6 +7,7 @@ open module com.swirlds.platform.core.test.fixtures {
     requires transitive com.swirlds.state.api.test.fixtures;
     requires transitive com.swirlds.state.api;
     requires transitive com.swirlds.virtualmap;
+    requires com.swirlds.base;
     requires com.swirlds.logging;
     requires com.swirlds.merkledb;
     requires com.hedera.pbj.runtime;
