+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/build.gradle.kts
@@ -31,7 +31,6 @@ testModuleInfo {
     requires("com.swirlds.base.test.fixtures")
     requires("awaitility")
     requires("org.junit.jupiter.params")
-    requires("org.assertj.core")
     requires("org.mockito.junit.jupiter")
     requires("com.swirlds.metrics.api")
     requires("org.mockito")
